+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+x.x.x Release notes (yyyy-MM-dd)
+=============================================================
+
+### API breaking changes
+
+* None.
+
+### Enhancements
+
+* None.
+
+### Bugfixes
+
+* Fix a crash when reading the shared schema from an observed Swift object.
+
 1.0.1 Release notes (2016-06-12)
 =============================================================